Add `RtcAudioEffectClient`
