+++ package.json
@@ -21,11 +21,11 @@
     "**/video.js": "7.10.0"
   },
   "dependencies": {
-    "@netless/app-slide": "0.2.44",
+    "@netless/app-slide": "0.2.47",
     "@netless/combine-player": "^1.1.1",
     "@netless/cursor-tool": "^0.1.1",
     "@netless/iframe-bridge": "2.1.9",
-    "@netless/slide-rtc-plugin": "0.10.1",
+    "@netless/slide-rtc-plugin": "1.1.9",
     "@netless/synced-store": "^2.0.3",
     "@netless/video-js-plugin": "0.3.8",
     "@netless/webview-bridge": "^0.1.9",
@@ -52,7 +52,7 @@
     "@babel/preset-env": "^7.7.4",
     "@babel/preset-react": "^7.7.4",
     "@babel/preset-typescript": "^7.7.4",
-    "@netless/whiteboard-bridge-types": "^0.1.11",
+    "@netless/whiteboard-bridge-types": "^0.1.12",
     "@types/css-font-loading-module": "^0.0.4",
     "@types/react": "^16.9.13",
     "@types/react-dom": "^16.9.4",
